Moved integration test to its package
